Rewriting Build_and_test_standards file
This is the main STDCI file. This rewrite updates it to our more recent
concepts about it, and prepares it for adding the many new features
we've added recently.

The main motivation behind this work is to enable splitting away the
Gerrit-Freestyle-JJB-YAML-Specific parts so the new GitHub oriented
features can be documented as well.

In more detail, this patch splits the old document into four parts:

1. A main part that explains the key concepts behind standard-CI.
2. A separate document about how to install and use `mock_runner.sh`
3. A separate document about how to write JJB YAML to setup
   Freestyle/Gerrit STDCI jobs.
4. A new separate document about how to setup Pipeline/GitHub STDCI jobs

Generally most text had been expanded or rewritten in one way or another.

Change-Id: Ie0422c664039e9895da6ee119986f6835b7e146c
Signed-off-by: Barak Korren <bkorren@redhat.com>
Jira-Ticket: https://ovirt-jira.atlassian.net/browse/OVIRT-1548